LayoutTests/canvas/philip/tests/type.name-expected.txt

< [index] >
type.name
HTMLCanvasElement type and toString
References: canvas.type
Actual output:
Passed
